+++ src/treeiksolverpos_nr_jl.hpp
@@ -62,8 +62,7 @@ public:
     TreeIkSolverPos_NR_JL(const Tree& tree, const std::vector<std::string>& endpoints, const JntArray& q_min, const JntArray& q_max, TreeFkSolverPos& fksolver,TreeIkSolverVel& iksolver,unsigned int maxiter=100,double eps=1e-6);
     ~TreeIkSolverPos_NR_JL();
 
-    virtual int CartToJnt(const JntArray& q_init, const Frames& p_in,
-            JntArray& q_out);
+    virtual double CartToJnt(const JntArray& q_init, const Frames& p_in, JntArray& q_out);
 
 private:
     const Tree tree;
@@ -75,7 +74,7 @@ private:
     Frames frames;
     Twists delta_twists;
     std::vector<std::string> endpoints;
-
+    
     unsigned int maxiter;
     double eps;
 };
